backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Environment and configuration
python-dotenv==1.0.0

# HTTP client
httpx==0.26.0
requests==2.31.0

# OpenAI for AI evaluation
openai==1.10.0

# Utilities
python-multipart==0.0.6

# Existing dependencies
